+++ sdk/cosmos/azure_data_cosmos/src/feed.rs
@@ -7,12 +7,11 @@ use azure_core::http::{
     headers::Headers,
     pager::{PagerContinuation, PagerResult},
 };
-#[cfg(not(target_arch = "wasm32"))]
 use futures::stream::BoxStream;
 use futures::Stream;
 use serde::{de::DeserializeOwned, Deserialize};
 
-use crate::{conditional_send::ConditionalSend, constants, models::CosmosResponse};
+use crate::{constants, models::CosmosResponse};
 
 /// Represents a single page of results from a Cosmos DB feed.
 ///
@@ -116,20 +115,16 @@ impl<T: DeserializeOwned> FeedPage<T> {
 ///
 /// See [`FeedPage`] for more details on Cosmos DB feeds.
 #[pin_project::pin_project]
-pub struct FeedItemIterator<T: ConditionalSend> {
+pub struct FeedItemIterator<T: Send> {
     #[pin]
-    #[cfg(not(target_arch = "wasm32"))]
     pages: BoxStream<'static, azure_core::Result<FeedPage<T>>>,
-    #[pin]
-    #[cfg(target_arch = "wasm32")]
-    pages: Pin<Box<dyn Stream<Item = azure_core::Result<FeedPage<T>>> + 'static>>,
     current: Option<std::vec::IntoIter<T>>,
 }
 
-impl<T: ConditionalSend> FeedItemIterator<T> {
+impl<T: Send> FeedItemIterator<T> {
     /// Creates a new `FeedItemIterator` from a stream of pages.
     pub(crate) fn new(
-        stream: impl Stream<Item = azure_core::Result<FeedPage<T>>> + ConditionalSend + 'static,
+        stream: impl Stream<Item = azure_core::Result<FeedPage<T>>> + Send + 'static,
     ) -> Self {
         Self {
             pages: Box::pin(stream),
@@ -142,7 +137,7 @@ impl<T: ConditionalSend> FeedItemIterator<T> {
     }
 }
 
-impl<T: ConditionalSend> Stream for FeedItemIterator<T> {
+impl<T: Send> Stream for FeedItemIterator<T> {
     type Item = azure_core::Result<T>;
 
     fn poll_next(
@@ -175,13 +170,9 @@ impl<T: ConditionalSend> Stream for FeedItemIterator<T> {
     }
 }
 
-pub struct FeedPageIterator<T: ConditionalSend>(
-    #[cfg(not(target_arch = "wasm32"))] BoxStream<'static, azure_core::Result<FeedPage<T>>>,
-    #[cfg(target_arch = "wasm32")]
-    Pin<Box<dyn Stream<Item = azure_core::Result<FeedPage<T>>> + 'static>>,
-);
+pub struct FeedPageIterator<T: Send>(BoxStream<'static, azure_core::Result<FeedPage<T>>>);
 
-impl<T: ConditionalSend> Stream for FeedPageIterator<T> {
+impl<T: Send> Stream for FeedPageIterator<T> {
     type Item = azure_core::Result<FeedPage<T>>;
 
     fn poll_next(